refactor: moved all db models to core project
