+++ src/lori.backend.Core/Models/Customer.cs
@@ -1,4 +1,4 @@
-﻿namespace lori.backend.Infrastructure.Models;
+﻿namespace lori.backend.Core.Models;
 
 public class Customer
 {
