+++ src/lori.backend.Core/Models/Item.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace lori.backend.Infrastructure.Models;
+﻿namespace lori.backend.Core.Models;
 public class Item
 {
   public int Id { get; set; }
