+++ src/lori.backend.Core/Models/Login.cs
@@ -1,11 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.ComponentModel.DataAnnotations;
 
-namespace lori.backend.Infrastructure.Models;
+namespace lori.backend.Core.Models;
 public class Login
 {
   [Key]
