+++ src/lori.backend.Core/Models/Role.cs
@@ -1,0 +1,7 @@
+﻿namespace lori.backend.Core.Models;
+public class Role
+{
+  public int Id { get; set; }
+  public string Name { get; set; } = null!;
+  public string Description { get; set; } = null!;
+}
