+++ src/lori.backend.Core/Models/Store.cs
@@ -1,0 +1,7 @@
+﻿namespace lori.backend.Core.Models;
+public class Store
+{
+  public int Id { get; set; }
+  public string Name { get; set; } = null!;
+  public int Quantity { get; set; }
+}
