--- src/lori.backend.Infrastructure/Models/Role.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace lori.backend.Infrastructure.Models;
-public class Role
-{
-  public int Id { get; set; }
-  public string Name { get; set; } = null!;
-  public string Description { get; set; } = null!;
-}
